test-kata: update so we can handle in local and external repos
Before, we would apply YAML from PKG_SHA branch of the packaging
repository directly. This isn't easy to handle from forks, which will be
what is supplied to the action when being triggerd directly from
packaging repository.

Let's change the test to check for presence of the git repository (ie,
if called from packaging repo flow), and clone it locally if it doesn't exits (if its
called from kata-containers repo flow).

Signed-off-by: Eric Ernst <eric.ernst@intel.com>